+++ src/main/java/ch/ethz/idsc/amodeus/analysis/report/DistanceElementHtml.java
@@ -23,31 +23,29 @@ public class DistanceElementHtml implements HtmlReportElement {
         // Aggregated Results:
         String aRKey = BodyElementKeys.AGGREGATERESULTS;
         HtmlBodyElement aRElement = new HtmlBodyElement();
-        aRElement.getHTMLGenerator()
-                .insertTextLeft("\nDistance Ratio:" + //
-                        "\nOccupancy Ratio:" + //
-                        "\n" + //
-                        "\n" + aRElement.getHTMLGenerator().bold("Distances") + //
-                        "\n\tTotal:" + //
-                        "\n\tRebalancing:" + //
-                        "\n\tPickup:" + //
-                        "\n\tWith Customer:" + //
-                        "\n" + //
-                        "\nAverage Trip Distance:" //
+        aRElement.getHTMLGenerator().insertTextLeft("\nDistance Ratio:" + //
+                "\nOccupancy Ratio:" + //
+                "\n" + //
+                "\n" + aRElement.getHTMLGenerator().bold("Distances") + //
+                "\n\tTotal:" + //
+                "\n\tRebalancing:" + //
+                "\n\tPickup:" + //
+                "\n\tWith Customer:" + //
+                "\n" + //
+                "\nAverage Trip Distance:" //
         );
-        aRElement.getHTMLGenerator()
-                .insertTextLeft("\n" + DECIMAL.format(de.totalDistanceRatio * 100) + "%" + //
-                        "\n" + DECIMAL.format(Mean.of(de.occupancyTensor).Get().number().doubleValue() * 100) + " %" + //
-                        "\n\n" + //
-                        "\n" + DECIMAL.format(de.totalDistance) + " km" + //
-                        "\n" + DECIMAL.format(de.totalDistanceRebal) + " km (" + //
-                        DECIMAL.format(100 * de.totalDistanceRebal / de.totalDistance) + "%)" + //
-                        "\n" + DECIMAL.format(de.totalDistancePicku) + " km (" + //
-                        DECIMAL.format(100 * de.totalDistancePicku / de.totalDistance) + "%)" + //
-                        "\n" + DECIMAL.format(de.totalDistanceWtCst) + " km (" + //
-                        DECIMAL.format(100 * de.totalDistanceWtCst / de.totalDistance) + "%)" + //
-                        "\n" + //
-                        "\n" + DECIMAL.format(de.totalDistanceWtCst / de.requestIndices.size()) + " km");
+        aRElement.getHTMLGenerator().insertTextLeft("\n" + DECIMAL.format(de.totalDistanceRatio * 100) + "%" + //
+                "\n" + DECIMAL.format(Mean.of(de.occupancyTensor).Get().number().doubleValue() * 100) + " %" + //
+                "\n\n" + //
+                "\n" + DECIMAL.format(de.totalDistance) + " km" + //
+                "\n" + DECIMAL.format(de.totalDistanceRebal) + " km (" + //
+                DECIMAL.format(100 * de.totalDistanceRebal / de.totalDistance) + "%)" + //
+                "\n" + DECIMAL.format(de.totalDistancePicku) + " km (" + //
+                DECIMAL.format(100 * de.totalDistancePicku / de.totalDistance) + "%)" + //
+                "\n" + DECIMAL.format(de.totalDistanceWtCst) + " km (" + //
+                DECIMAL.format(100 * de.totalDistanceWtCst / de.totalDistance) + "%)" + //
+                "\n" + //
+                "\n" + DECIMAL.format(de.totalDistanceWtCst / de.requestIndices.size()) + " km");
         File img = new File(IMAGE_FOLDER, StackedDistanceChartImage.FILENAME + ".png");
         aRElement.getHTMLGenerator().insertImg(img.getPath(), 700, 125);
         bodyElements.put(aRKey, aRElement);
