+++ src/main/java/ch/ethz/idsc/amodeus/dispatcher/distance/AStarEuclideanFactory.java
@@ -1,4 +1,4 @@
-package ch.ethz.idsc.amodeus.dispatcher.util.distance_function;
+package ch.ethz.idsc.amodeus.dispatcher.distance;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
@@ -7,12 +7,12 @@ import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.network.Network;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.core.router.ArrayFastRouterDelegateFactory;
-import org.matsim.core.router.FastAStarLandmarks;
+import org.matsim.core.router.FastAStarEuclidean;
 import org.matsim.core.router.FastRouterDelegateFactory;
 import org.matsim.core.router.util.ArrayRoutingNetworkFactory;
 import org.matsim.core.router.util.LeastCostPathCalculator;
 import org.matsim.core.router.util.LeastCostPathCalculatorFactory;
-import org.matsim.core.router.util.PreProcessLandmarks;
+import org.matsim.core.router.util.PreProcessEuclidean;
 import org.matsim.core.router.util.RoutingNetwork;
 import org.matsim.core.router.util.RoutingNetworkFactory;
 import org.matsim.core.router.util.RoutingNetworkNode;
@@ -20,18 +20,17 @@ import org.matsim.core.router.util.TravelDisutility;
 import org.matsim.core.router.util.TravelTime;
 import org.matsim.vehicles.Vehicle;
 
-public class AStarLandmarksFactory implements LeastCostPathCalculatorFactory {
+/** @author sebhoerl */
+public class AStarEuclideanFactory implements LeastCostPathCalculatorFactory {
     final private RoutingNetworkFactory routingNetworkFactory = new ArrayRoutingNetworkFactory();
-    final private int numberOfLandmarks;
     final private double overdoFactor;
 
     private Network network;
 
     private RoutingNetwork routingNetwork;
-    private PreProcessLandmarks preProcessLandmarks;
+    private PreProcessEuclidean preProcessEucledian;
 
-    public AStarLandmarksFactory(int numberOfLandmarks, double overdofactor) {
-        this.numberOfLandmarks = numberOfLandmarks;
+    public AStarEuclideanFactory(double overdofactor) {
         this.overdoFactor = overdofactor;
     }
 
@@ -54,12 +53,11 @@ public class AStarLandmarksFactory implements LeastCostPathCalculatorFactory {
                 }
             };
 
-            preProcessLandmarks = new PreProcessLandmarks(disutility, numberOfLandmarks);
-            preProcessLandmarks.setNumberOfThreads(Runtime.getRuntime().availableProcessors());
-            preProcessLandmarks.run(network);
+            preProcessEucledian = new PreProcessEuclidean(disutility);
+            preProcessEucledian.run(network);
 
             for (RoutingNetworkNode node : routingNetwork.getNodes().values()) {
-                node.setDeadEndData(preProcessLandmarks.getNodeData(node.getNode()));
+                node.setDeadEndData(preProcessEucledian.getNodeData(node.getNode()));
             }
         } else if (!this.network.equals(network)) {
             throw new IllegalStateException();
@@ -68,11 +66,11 @@ public class AStarLandmarksFactory implements LeastCostPathCalculatorFactory {
         FastRouterDelegateFactory fastRouterFactory = new ArrayFastRouterDelegateFactory();
 
         try {
-            Constructor<FastAStarLandmarks> constructor = FastAStarLandmarks.class.getDeclaredConstructor(RoutingNetwork.class, PreProcessLandmarks.class, TravelDisutility.class,
+            Constructor<FastAStarEuclidean> constructor = FastAStarEuclidean.class.getDeclaredConstructor(RoutingNetwork.class, PreProcessEuclidean.class, TravelDisutility.class,
                     TravelTime.class, double.class, FastRouterDelegateFactory.class);
 
             constructor.setAccessible(true);
-            return constructor.newInstance(routingNetwork, preProcessLandmarks, travelCosts, travelTimes, overdoFactor, fastRouterFactory);
+            return constructor.newInstance(routingNetwork, preProcessEucledian, travelCosts, travelTimes, overdoFactor, fastRouterFactory);
         } catch (NoSuchMethodException | SecurityException | InstantiationException | IllegalAccessException | IllegalArgumentException | InvocationTargetException e) {
             throw new RuntimeException(e);
         }
