+++ src/main/java/ch/ethz/idsc/amodeus/dispatcher/distance/AStarLandmarksFactory.java
@@ -1,4 +1,4 @@
-package ch.ethz.idsc.amodeus.dispatcher.util.distance_function;
+package ch.ethz.idsc.amodeus.dispatcher.distance;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
@@ -7,12 +7,11 @@ import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.network.Network;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.core.router.ArrayFastRouterDelegateFactory;
-import org.matsim.core.router.FastAStarEuclidean;
+import org.matsim.core.router.FastAStarLandmarks;
 import org.matsim.core.router.FastRouterDelegateFactory;
 import org.matsim.core.router.util.ArrayRoutingNetworkFactory;
 import org.matsim.core.router.util.LeastCostPathCalculator;
 import org.matsim.core.router.util.LeastCostPathCalculatorFactory;
-import org.matsim.core.router.util.PreProcessEuclidean;
 import org.matsim.core.router.util.PreProcessLandmarks;
 import org.matsim.core.router.util.RoutingNetwork;
 import org.matsim.core.router.util.RoutingNetworkFactory;
@@ -21,16 +20,19 @@ import org.matsim.core.router.util.TravelDisutility;
 import org.matsim.core.router.util.TravelTime;
 import org.matsim.vehicles.Vehicle;
 
-public class AStarEuclideanFactory implements LeastCostPathCalculatorFactory {
-    final private RoutingNetworkFactory routingNetworkFactory = new ArrayRoutingNetworkFactory();
-    final private double overdoFactor;
+/** @author sebhoerl */
+public class AStarLandmarksFactory implements LeastCostPathCalculatorFactory {
+    private final RoutingNetworkFactory routingNetworkFactory = new ArrayRoutingNetworkFactory();
+    private final int numberOfLandmarks;
+    private final double overdoFactor;
 
     private Network network;
 
     private RoutingNetwork routingNetwork;
-    private PreProcessEuclidean preProcessEucledian;
+    private PreProcessLandmarks preProcessLandmarks;
 
-    public AStarEuclideanFactory(double overdofactor) {
+    public AStarLandmarksFactory(int numberOfLandmarks, double overdofactor) {
+        this.numberOfLandmarks = numberOfLandmarks;
         this.overdoFactor = overdofactor;
     }
 
@@ -53,11 +55,12 @@ public class AStarEuclideanFactory implements LeastCostPathCalculatorFactory {
                 }
             };
 
-            preProcessEucledian = new PreProcessEuclidean(disutility);
-            preProcessEucledian.run(network);
+            preProcessLandmarks = new PreProcessLandmarks(disutility, numberOfLandmarks);
+            preProcessLandmarks.setNumberOfThreads(Runtime.getRuntime().availableProcessors());
+            preProcessLandmarks.run(network);
 
             for (RoutingNetworkNode node : routingNetwork.getNodes().values()) {
-                node.setDeadEndData(preProcessEucledian.getNodeData(node.getNode()));
+                node.setDeadEndData(preProcessLandmarks.getNodeData(node.getNode()));
             }
         } else if (!this.network.equals(network)) {
             throw new IllegalStateException();
@@ -66,11 +69,11 @@ public class AStarEuclideanFactory implements LeastCostPathCalculatorFactory {
         FastRouterDelegateFactory fastRouterFactory = new ArrayFastRouterDelegateFactory();
 
         try {
-            Constructor<FastAStarEuclidean> constructor = FastAStarEuclidean.class.getDeclaredConstructor(RoutingNetwork.class, PreProcessEuclidean.class, TravelDisutility.class,
+            Constructor<FastAStarLandmarks> constructor = FastAStarLandmarks.class.getDeclaredConstructor(RoutingNetwork.class, PreProcessLandmarks.class, TravelDisutility.class,
                     TravelTime.class, double.class, FastRouterDelegateFactory.class);
 
             constructor.setAccessible(true);
-            return constructor.newInstance(routingNetwork, preProcessEucledian, travelCosts, travelTimes, overdoFactor, fastRouterFactory);
+            return constructor.newInstance(routingNetwork, preProcessLandmarks, travelCosts, travelTimes, overdoFactor, fastRouterFactory);
         } catch (NoSuchMethodException | SecurityException | InstantiationException | IllegalAccessException | IllegalArgumentException | InvocationTargetException e) {
             throw new RuntimeException(e);
         }
