+++ src/main/java/ch/ethz/idsc/amodeus/dispatcher/distance/DijkstraDistanceFunctionFactory.java
@@ -1,8 +1,9 @@
-package ch.ethz.idsc.amodeus.dispatcher.util.distance_function;
+package ch.ethz.idsc.amodeus.dispatcher.distance;
 
 import org.matsim.api.core.v01.network.Network;
 import org.matsim.core.router.FastDijkstraFactory;
 
+/** @author sebhoerl */
 public class DijkstraDistanceFunctionFactory implements DistanceFunctionFactory {
     @Override
     public DistanceFunction createDistanceFunction(Network network) {
