+++ src/main/java/ch/ethz/idsc/amodeus/dispatcher/distance/EuclideanDistanceFunctionFactory.java
@@ -1,7 +1,8 @@
-package ch.ethz.idsc.amodeus.dispatcher.util.distance_function;
+package ch.ethz.idsc.amodeus.dispatcher.distance;
 
 import org.matsim.api.core.v01.network.Network;
 
+/** @author sebhoerl */
 public class EuclideanDistanceFunctionFactory implements DistanceFunctionFactory {
     @Override
     public DistanceFunction createDistanceFunction(Network network) {
