+++ src/main/java/ch/ethz/idsc/amodeus/dispatcher/util/BipartiteMatchingUtils.java
@@ -10,8 +10,8 @@ import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.network.Network;
 
 import ch.ethz.idsc.amodeus.dispatcher.core.RoboTaxi;
-import ch.ethz.idsc.amodeus.dispatcher.util.distance_function.DistanceFunction;
-import ch.ethz.idsc.amodeus.dispatcher.util.distance_function.EuclideanDistanceFunction;
+import ch.ethz.idsc.amodeus.dispatcher.distance.DistanceFunction;
+import ch.ethz.idsc.amodeus.dispatcher.distance.EuclideanDistanceFunction;
 import ch.ethz.idsc.amodeus.util.math.GlobalAssert;
 import ch.ethz.idsc.tensor.Tensor;
 import ch.ethz.idsc.tensor.Tensors;
