+++ src/main/java/ch/ethz/idsc/amodeus/matsim/mod/AmodeusDriveTaskTracker.java
@@ -69,7 +69,7 @@ public class AmodeusDriveTaskTracker implements OnlineDriveTaskTracker {
      * next link, or</li>
      * <li>(b) no diversion possible (the leg ends on the current link)</li>
      * </ul>
-    */
+     */
     @Override
     public LinkTimePair getDiversionPoint() {
         if (vrpDynLeg.canChangeNextLink()) {
