+++ src/test/java/ch/ethz/idsc/amodeus/dispatcher/distance/DistanceFunctionTest.java
@@ -1,5 +1,5 @@
 /* amodeus - Copyright (c) 2018, ETH Zurich, Institute for Dynamic Systems and Control */
-package ch.ethz.idsc.amodeus.dispatcher.utils.distance_function;
+package ch.ethz.idsc.amodeus.dispatcher.distance;
 
 import java.util.Arrays;
 import java.util.Collection;
